pyproject.toml
[tool.poetry]
name = "stack"
version = "0.1.0"
description = ""
authors = ["Stephen <stephen@theyoungs.cloud>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyexiftool = "^0.5.6"
numpy = "^1.26.3"
pillow = "^10.2.0"
colorama = "^0.4.6"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
